+++ apps/docs/app/docs/[...slug]/page.tsx
@@ -107,7 +107,5 @@ export function generateMetadata({ params }: { params: Param }): Metadata {
 }
 
 export function generateStaticParams(): Param[] {
-  return utils.getPages().map((page) => ({
-    slug: page.slugs,
-  }));
+  return utils.generateParams();
 }