+++ examples/i18n/app/api/search/route.ts
@@ -1,10 +1,10 @@
-import { getLanguages } from '@/app/source';
+import { source } from '@/app/source';
 import { i18n } from '@/i18n';
 import { createI18nSearchAPIExperimental } from 'fumadocs-core/search/server';
 
 export const { GET } = createI18nSearchAPIExperimental('advanced', {
   i18n,
-  indexes: getLanguages().flatMap((entry) => {
+  indexes: source.getLanguages().flatMap((entry) => {
     return entry.pages.map((page) => ({
       id: page.url,
       url: page.url,