+++ examples/i18n/app/source.ts
@@ -3,7 +3,7 @@ import { loader } from 'fumadocs-core/source';
 import { i18n } from '@/i18n';
 import { docs, meta } from '@/.source';
 
-export const { getPage, getPages, getLanguages, pageTree } = loader({
+export const source = loader({
   baseUrl: '/',
   source: createMDXSource(docs, meta),
   i18n,