+++ examples/next-mdx/app/layout.config.tsx
@@ -1,0 +1,7 @@
+import { HomeLayoutProps } from 'fumadocs-ui/home-layout';
+
+export const baseOptions: HomeLayoutProps = {
+  nav: {
+    title: 'My App',
+  },
+};