+++ packages/create-app/template/fuma-docs-mdx/app/source.ts
@@ -2,7 +2,7 @@ import { docs, meta } from '@/.source';
 import { createMDXSource } from 'fumadocs-mdx';
 import { loader } from 'fumadocs-core/source';
 
-export const { getPage, getPages, pageTree } = loader({
+export const source = loader({
   baseUrl: '/docs',
   source: createMDXSource(docs, meta),
 });