+++ packages/ui/src/layout.shared.tsx
@@ -19,6 +19,13 @@ export interface SharedNavProps extends TitleProps, NavBoxProps {
 
 export interface BaseLayoutProps {
   /**
+   * Enable Language Switch
+   *
+   * @defaultValue false
+   */
+  i18n?: boolean;
+
+  /**
    * GitHub url
    */
   githubUrl?: string;